Added more indexes to DatasetVersion and DataTuple tables
Added exception handling to the primary indexing loginc and UI

Added exception handling to dataset sync methods and UI

Applied a check for the size of MV after refreshing it, throwing exception if not matched with the number of tunples.

Reports errors when sync fails
Reports errors when Reindex primary data fails

Tested on AquaDiva data. many of the datasets contain faulty data!